requirements.txt
deepagents
langchain-openai
PyYAML
